Add call to pokeapi to retrieve Pokemon description
This is a work in progress (WIP), the pokeapi interaction is not
finalized. One of the main things missing is proper handling
of situations outside of the happy path, such as:
- pokemon not found
- connectivity issues
- unexpected status codes from pokeapi

This commit also adds code to mock pokeapi in end-to-end tests,
though this is a work in progress too.