Add call to shakespeare translation API to translate Pokemon description
This is a work in progress (WIP), the shakespeare API interaction is not
finalized. One of the main things missing is proper handling
of situations outside of the happy path, such as:
- rate-limit error
- connectivity issues
- unexpected status codes from pokeapi

This commit also adds code to mock shakespeare API in end-to-end tests,
though this is a work in progress too.