Handle some of the possible errors which can happen while querying the pokeapi endpoint
This commit introduces a custom internal error which maps automatically to
Actix errors (see https://github.com/actix/examples/blob/master/error_handling/src/main.rs).
The custom error is named CustomError for now but will probably be renamed.

An integration test to check the behavior when the pokemon is not
found has been added (the service correctly returns 404 in this case).