Centralize handling of errors coming from external service interaction
This is only a refactor which does not add additional functionality.
The error handling (and related logging) was making functions difficult to
read, so we enrich our error model to capture more information
in order to push logging in one centralized layer.

With this change, the application logs errors when the error is
translated into an actix_web response error.

This is a tradeoff, the functions are more readable but error handling
is hidden from the main flow.